+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2026-02-nn [upcoming] ##
-* Aarch64 images with uefi loader.
-  * Set "platform: aarch64-uefi" in tarbsd.yml.
+* Aarch64 support.
+  * Set "platform: aarch64" in tarbsd.yml.
   * Likely wont't boot on every random board there is.
 * Package install step has gotten a speed improvement.
 * Work file system will grow automatically to accomodate building bigger images, wrk-init